stable! more design needs to be implemented though

+++ src/henrybryan/friendquiz/business/Calculate.java
@@ -9,7 +9,7 @@ package henrybryan.friendquiz.business;
  */
 public class Calculate {
 	
-	public static double claculatePercentage(int totalScore, int maxScore) {
+	public static double claculatePercentage(double totalScore, int maxScore) {
 		double percent = 0.0;
 		percent = (double)Math.round(((totalScore/maxScore)*100.0)*100)/100;
 		return percent;
